+++ objc2-foundation/src/mutable_string.rs
@@ -1,0 +1,211 @@
+use core::cmp;
+use core::fmt;
+use core::ops::AddAssign;
+use core::ptr::NonNull;
+use core::str;
+
+use objc2::msg_send;
+use objc2::rc::{DefaultId, Id, Owned, Shared};
+
+use crate::{NSCopying, NSMutableCopying, NSString};
+
+object! {
+    /// A dynamic plain-text Unicode string object.
+    ///
+    /// See [Apple's documentation](https://developer.apple.com/documentation/foundation/nsmutablestring?language=objc).
+    unsafe pub struct NSMutableString: NSString;
+}
+
+// TODO: SAFETY
+unsafe impl Sync for NSMutableString {}
+unsafe impl Send for NSMutableString {}
+
+/// Creating mutable strings.
+impl NSMutableString {
+    unsafe_def_fn! {
+        /// Construct an empty [`NSMutableString`].
+        pub fn new -> Owned;
+    }
+
+    /// Creates a new [`NSMutableString`] by copying the given string slice.
+    #[doc(alias = "initWithBytes:length:encoding:")]
+    pub fn from_str(string: &str) -> Id<Self, Owned> {
+        unsafe {
+            let obj = super::string::from_str(Self::class(), string);
+            Id::new(obj.cast())
+        }
+    }
+
+    /// Creates a new [`NSMutableString`] from the given [`NSString`].
+    #[doc(alias = "initWithString:")]
+    pub fn from_nsstring(string: &NSString) -> Id<Self, Owned> {
+        unsafe {
+            let obj: *mut Self = msg_send![Self::class(), alloc];
+            let obj: *mut Self = msg_send![obj, initWithString: string];
+            Id::new(NonNull::new_unchecked(obj))
+        }
+    }
+
+    #[doc(alias = "initWithCapacity:")]
+    pub fn with_capacity(capacity: usize) -> Id<Self, Owned> {
+        unsafe {
+            let obj: *mut Self = msg_send![Self::class(), alloc];
+            let obj: *mut Self = msg_send![obj, initWithCapacity: capacity];
+            Id::new(NonNull::new_unchecked(obj))
+        }
+    }
+}
+
+/// Mutating strings.
+impl NSMutableString {
+    /// Appends the given [`NSString`] onto the end of this.
+    #[doc(alias = "appendString:")]
+    pub fn push_nsstring(&mut self, nsstring: &NSString) {
+        // SAFETY: The string is not nil
+        unsafe { msg_send![self, appendString: nsstring] }
+    }
+
+    /// Replaces the entire string.
+    #[doc(alias = "setString:")]
+    pub fn replace(&mut self, nsstring: &NSString) {
+        // SAFETY: The string is not nil
+        unsafe { msg_send![self, setString: nsstring] }
+    }
+
+    // TODO:
+    // - deleteCharactersInRange:
+    // - replaceCharactersInRange:withString:
+    // - insertString:atIndex:
+    // Figure out how these work on character boundaries
+}
+
+impl DefaultId for NSMutableString {
+    type Ownership = Owned;
+
+    #[inline]
+    fn default_id() -> Id<Self, Self::Ownership> {
+        Self::new()
+    }
+}
+
+unsafe impl NSCopying for NSMutableString {
+    type Ownership = Shared;
+    type Output = NSString;
+}
+
+unsafe impl NSMutableCopying for NSMutableString {
+    type Output = NSMutableString;
+}
+
+impl AddAssign<&NSString> for NSMutableString {
+    #[inline]
+    fn add_assign(&mut self, other: &NSString) {
+        self.push_nsstring(other)
+    }
+}
+
+impl PartialEq<NSString> for NSMutableString {
+    #[inline]
+    fn eq(&self, other: &NSString) -> bool {
+        PartialEq::eq(&**self, other)
+    }
+}
+
+impl PartialEq<NSMutableString> for NSString {
+    #[inline]
+    fn eq(&self, other: &NSMutableString) -> bool {
+        PartialEq::eq(self, &**other)
+    }
+}
+
+impl PartialOrd for NSMutableString {
+    #[inline]
+    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
+        PartialOrd::partial_cmp(&**self, &**other)
+    }
+}
+
+impl PartialOrd<NSString> for NSMutableString {
+    #[inline]
+    fn partial_cmp(&self, other: &NSString) -> Option<cmp::Ordering> {
+        PartialOrd::partial_cmp(&**self, other)
+    }
+}
+
+impl PartialOrd<NSMutableString> for NSString {
+    #[inline]
+    fn partial_cmp(&self, other: &NSMutableString) -> Option<cmp::Ordering> {
+        PartialOrd::partial_cmp(self, &**other)
+    }
+}
+
+impl Ord for NSMutableString {
+    #[inline]
+    fn cmp(&self, other: &Self) -> cmp::Ordering {
+        Ord::cmp(&**self, &**other)
+    }
+}
+
+impl fmt::Write for NSMutableString {
+    fn write_str(&mut self, s: &str) -> Result<(), fmt::Error> {
+        let nsstring = NSString::from_str(s);
+        self.push_nsstring(&nsstring);
+        Ok(())
+    }
+}
+
+impl fmt::Display for NSMutableString {
+    #[inline]
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::Display::fmt(&**self, f)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use alloc::string::ToString;
+
+    use super::*;
+
+    #[test]
+    fn test_from_nsstring() {
+        let s = NSString::from_str("abc");
+        let s = NSMutableString::from_nsstring(&s);
+        assert_eq!(&s.to_string(), "abc");
+    }
+
+    #[test]
+    fn test_push_nsstring() {
+        let mut s = NSMutableString::from_str("abc");
+        s.push_nsstring(&NSString::from_str("def"));
+        *s += &NSString::from_str("ghi");
+        assert_eq!(&s.to_string(), "abcdefghi");
+    }
+
+    #[test]
+    fn test_replace() {
+        let mut s = NSMutableString::from_str("abc");
+        s.replace(&NSString::from_str("def"));
+        assert_eq!(&s.to_string(), "def");
+    }
+
+    #[test]
+    fn test_with_capacity() {
+        let mut s = NSMutableString::with_capacity(3);
+        *s += &NSString::from_str("abc");
+        *s += &NSString::from_str("def");
+        assert_eq!(&s.to_string(), "abcdef");
+    }
+
+    #[test]
+    fn test_copy() {
+        let s1 = NSMutableString::from_str("abc");
+        let s2 = s1.copy();
+        assert_ne!(s1.as_ptr(), s2.as_ptr() as *mut NSMutableString);
+        assert!(s2.is_kind_of(NSString::class()));
+
+        let s3 = s1.mutable_copy();
+        assert_ne!(s1.as_ptr(), s3.as_ptr());
+        assert!(s3.is_kind_of(NSMutableString::class()));
+    }
+}